+++ ethereum/contracts/icco/contributor/ContributorStructs.sol
@@ -26,7 +26,7 @@ contract ContributorStructs {
         // solidity does not handle struct arrays in storage well
         uint16[] acceptedTokensChains;
         bytes32[] acceptedTokensAddresses;
-        uint256[] acceptedTokensConversionRates;
+        uint128[] acceptedTokensConversionRates;
 
         // recipient of proceeds
         bytes32 recipient;
